+++ changelog.md
@@ -1,3 +1,7 @@
+# Version 2023.4.0 (2023-04-04)
+
+- Update requirements
+
 # Version 2023.3.1 (2023-03-17)
 
 - Add name to tasks